review comments, better synergy with examples
